LaraAdmin Install Complete

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -1,8 +1,13 @@
 <?php
+/**
+ * Migration genrated using LaraAdmin
+ * Help: http://laraadmin.com
+ */
 
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Dwij\Laraadmin\Models\Module;
 
 class CreateUsersTable extends Migration
 {
@@ -13,14 +18,48 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->string('password');
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        Module::generate("Users", 'users', 'name', 'fa-group', [
+            ["name", "Name", "Name", false, "", 5, 250, true],
+            ["context_id", "Context", "Integer", false, "0", 0, 0, false],
+            ["email", "Email", "Email", true, "", 0, 250, false],
+            ["password", "Password", "Password", false, "", 6, 250, true],
+            ["type", "User Type", "Dropdown", false, "Employee", 0, 0, false, ["Employee", "Client"]],
+        ]);
+		
+		/*
+		Row Format:
+		["field_name_db", "Label", "UI Type", "Unique", "Default_Value", "min_length", "max_length", "Required", "Pop_values"]
+        Module::generate("Module_Name", "Table_Name", "view_column_name" "Fields_Array");
+        
+		Module::generate("Books", 'books', 'name', [
+            ["address",     "Address",      "Address",  false, "",          0,  1000,   true],
+            ["restricted",  "Restricted",   "Checkbox", false, false,       0,  0,      false],
+            ["price",       "Price",        "Currency", false, 0.0,         0,  0,      true],
+            ["date_release", "Date of Release", "Date", false, "NULL", 0, 0,   false],
+            ["time_started", "Start Time",  "Datetime", false, "NOW()", 0, 0, false],
+            ["weight",      "Weight",       "Decimal",  false, 0.0,         0,  20,     true],
+            ["publisher",   "Publisher",    "Dropdown", false, "Marvel",    0,  0,      false, ["Bloomsbury","Marvel","Universal"]],
+            ["publisher",   "Publisher",    "Dropdown", false, 3,           0,  0,      false, "@publishers"],
+            ["email",       "Email",        "Email",    false, "",          0,  0,      false],
+            ["file",        "File",         "File",     false, "",          0,  1,      false],
+            ["files",       "Files",        "Files",    false, "",          0,  10,     false],
+            ["weight",      "Weight",       "Float",    false, 0.0,         0,  20.00,  true],
+            ["biography",   "Biography",    "HTML",     false, "<p>This is description</p>", 0, 0, true],
+            ["profile_image", "Profile Image", "Image", false, "img_path.jpg", 0, 250,  false],
+            ["pages",       "Pages",        "Integer",  false, 0,           0,  5000,   false],
+            ["mobile",      "Mobile",       "Mobile",   false, "+91  8888888888", 0, 20,false],
+            ["media_type",  "Media Type",   "Multiselect", false, ["Audiobook"], 0, 0,  false, ["Print","Audiobook","E-book"]],
+            ["media_type",  "Media Type",   "Multiselect", false, [2,3],    0,  0,      false, "@media_types"],
+            ["name",        "Name",         "Name",     false, "John Doe",  5,  250,    true],
+            ["password",    "Password",     "Password", false, "",          6,  250,    true],
+            ["status",      "Status",       "Radio",    false, "Published", 0,  0,      false, ["Draft","Published","Unpublished"]],
+            ["author",      "Author",       "String",   false, "JRR Tolkien", 0, 250,   true],
+            ["genre",       "Genre",        "Taginput", false, ["Fantacy","Adventure"], 0, 0, false],
+            ["description", "Description",  "Textarea", false, "",          0,  1000,   false],
+            ["short_intro", "Introduction", "TextField",false, "",          5,  250,    true],
+            ["website",     "Website",      "URL",      false, "http://dwij.in", 0, 0,  false],
+        ]);
+		*/
     }
 
     /**
@@ -30,6 +69,8 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('users');
+        if (Schema::hasTable('users')) {
+            Schema::drop('users');
+        }
     }
 }
